requirements.txt
kafka-python==2.0.2
numpy==1.23.0
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
